+++ packages/cli/oas/oas-github-ci/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @medusajs/oas-github-ci
 
+## 2.1.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @medusajs/medusa-oas-cli@2.1.0
+
 ## 2.0.7
 
 ### Patch Changes