+++ packages/modules/cache-inmemory/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @medusajs/cache-inmemory
 
+## 2.1.0
+
 ## 2.0.7
 
 ## 2.0.6